2-1

+++ goods/src/main/java/cn/edu/xmu/oomall/goods/model/vo/NewOnSaleVo.java
@@ -32,7 +32,7 @@ public class NewOnSaleVo {
     private Integer quantity;
 
     @NotNull
-    private Integer type;
+    private Byte type;
 
     public NewOnSaleVo(){}
